fix for rust 1.0-alpha

+++ src/matchers/matcher.rs
@@ -6,6 +6,6 @@ pub trait Matcher <T: Show> {
     fn assert_check(&self, expected: T) -> bool;
     fn msg(&self, expected: T) -> String;
     fn negated_msg(&self, expected: T) -> String;
-    fn get_file_line(&self) -> (&'static str, uint);
+    fn get_file_line(&self) -> (&'static str, usize);
 }
